feat: fix glitch in auth0
This commit adds a hacky workaround for the glitch that was occuring with the auth0 where the homepage (a non protected route) wouldn't render if a user was logged in and accessed it directly through the url (starting a new session). Might need more work but it's usable for now, but not ideal.